requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
prophet
pydantic
python-dotenv
redis
requests
